+++ test/test_expression.py
@@ -254,7 +254,7 @@ class TestExpression(unittest.TestCase):
         )
 
     def test_unknown_expression_parts(self):
-        line_id = LineIdentifier(1927, 'test_character_ordinals_in_expressions')
+        line_id = LineIdentifier(1928, 'test_unknown_expression_parts')
 
         with self.assertRaises(SystemExit, msg='extraneous comparison operator'):
             parse_expression(line_id, '<$2024').get_value(TestExpression.label_values, 1)